move OS structures
